Refactor entropy coding module
